+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [4.19.1](https://github.com/contentful/ui-extensions-sdk/compare/v4.19.0...v4.19.1) (2023-06-15)
+
+### Bug Fixes
+
+- [] Add boolean field type ([#1541](https://github.com/contentful/ui-extensions-sdk/issues/1541)) ([ba52bdf](https://github.com/contentful/ui-extensions-sdk/commit/ba52bdfc5a100348bc576ab6d0b13cfe17b36d47))
+
+# [4.19.0](https://github.com/contentful/ui-extensions-sdk/compare/v4.18.0...v4.19.0) (2023-06-14)
+
+### Features
+
+- [EXT-3648][EXT-4380] Improve sdk.field types ([#1529](https://github.com/contentful/ui-extensions-sdk/issues/1529)) ([ec0da5a](https://github.com/contentful/ui-extensions-sdk/commit/ec0da5a2ed399c18b1a479bcaa193cc480172c71))
+
 # [4.18.0](https://github.com/contentful/ui-extensions-sdk/compare/v4.17.2...v4.18.0) (2023-06-14)
 
 ### Features